Add an extraction for the XLSX available from https://www.isobus.net/isobus/
For some reason the ISO 11783 standard has some stuff provided to the
public at https://www.isobus.net/isobus/, and it also uses XLS to
disseminate, but instead of the DigitalAnnexes....it uses multiple XLSX
workbooks. (It's very strange these bodies are publishing "specs" as XL
files, not as XML or JSON or something machine readable).