Ensure PGN record is looked up or created for the SPN
The code previously assumed the processing was done in PGN-sorted order,
thus `current_pgn` would be correct for a given row. However, the
assumption about sort order is unsafe (because it can cause PGN records
to be over-written accidentally), and is unnecessary.

@timsouth94's changes here https://github.com/timsouth94/decoda/commit/8ee54bcec13dfcde041cf28f9743e6fa43508236
solve this. These changes do something very similar by making the
intermediate `result` value a dict by id.